refactor: refactor api for manage time-slot
